Rename middleware to http
