+++ pkg/http/http.go
@@ -1,4 +1,4 @@
-package middleware
+package http
 
 import (
 	"bufio"
